refactor: move JazzClient tests into integration tests dir
