+++ crates/jazz-tools/src/query_manager/rebac_tests/inherited_policies.rs
@@ -1,6 +1,3 @@
-#[cfg(feature = "test-utils")]
-use crate::JazzClient;
-
 use super::*;
 
 #[test]
@@ -270,253 +267,6 @@ fn rebac_inherited_insert_uses_requested_branch_instead_of_reusing_cached_branch
         "Document insert should apply once the requested parent branch is consulted"
     );
 }
-
-#[cfg(feature = "test-utils")]
-#[tokio::test]
-async fn rebac_inherits_filters_select_query_results() {
-    // Schema with INHERITS policy
-    let folders_table = TableSchema::builder("folders")
-        .column("owner_id", ColumnType::Text)
-        .column("name", ColumnType::Text);
-    let folders_policies = permissions(|p| {
-        p.allow_read()
-            .where_(pe::eq("owner_id", pe::session("user_id")));
-    });
-
-    let docs_table = TableSchema::builder("documents")
-        .column("owner_id", ColumnType::Text)
-        .column("title", ColumnType::Text)
-        .nullable_fk_column("folder_id", "folders");
-
-    // SELECT policy: owner_id = @user_id OR INHERITS SELECT VIA folder_id
-    let docs_policies = permissions(|p| {
-        p.allow_read().where_(pe::any_of([
-            pe::eq("owner_id", pe::session("user_id")),
-            pe::allowed_to_read("folder_id"),
-        ]));
-    });
-    let schema = SchemaBuilder::new()
-        .table(folders_table.policies(folders_policies))
-        .table(docs_table.policies(docs_policies))
-        .build();
-    let client = JazzClient::test_client(schema).await;
-
-    let folder_id = client
-        .insert(
-            "folders",
-            crate::row_input!("owner_id" => "alice", "name" => "Alice's Folder"),
-        )
-        .expect("seed folder")
-        .0;
-    client
-        .insert(
-            "documents",
-            crate::row_input!(
-                "owner_id" => "bob",
-                "title" => "Bob's Doc in Alice's Folder",
-                "folder_id" => folder_id,
-            ),
-        )
-        .expect("seed document");
-
-    let charlie_rows = client
-        .for_session(Session::new("charlie"))
-        .query(QueryBuilder::new("documents").build(), None)
-        .await
-        .expect("query documents as charlie");
-
-    // Charlie should NOT see Bob's document (doesn't own it, doesn't own folder)
-    assert!(
-        charlie_rows.is_empty(),
-        "Charlie should not see Bob's document - he owns neither the doc nor the folder. \
-         INHERITS should have denied access, but currently it always returns true."
-    );
-}
-
-#[cfg(feature = "test-utils")]
-#[tokio::test]
-async fn inherits_select_denies_when_parent_operation_policy_is_missing() {
-    let documents_policies = permissions(|p| {
-        p.allow_read().where_(pe::allowed_to_read("folder_id"));
-    });
-    let schema = SchemaBuilder::new()
-        .table(
-            TableSchema::builder("folders")
-                .column("owner_id", ColumnType::Text)
-                .column("name", ColumnType::Text),
-        )
-        .table(
-            TableSchema::builder("documents")
-                .column("owner_id", ColumnType::Text)
-                .column("title", ColumnType::Text)
-                .nullable_fk_column("folder_id", "folders")
-                .policies(documents_policies),
-        )
-        .build();
-    let client = JazzClient::test_client(schema).await;
-
-    let folder = client
-        .insert(
-            "folders",
-            crate::row_input!("owner_id" => "alice", "name" => "Shared"),
-        )
-        .expect("folder insert should succeed");
-    client
-        .insert(
-            "documents",
-            crate::row_input!("owner_id" => "bob", "title" => "Inherited doc", "folder_id" => folder.0))
-        .expect("document insert should succeed");
-
-    let rows = client
-        .for_session(Session::new("alice"))
-        .query(
-            QueryBuilder::new("documents").select(&["title"]).build(),
-            None,
-        )
-        .await
-        .expect("query documents as alice");
-
-    assert!(
-        rows.is_empty(),
-        "child rows should be denied when INHERITS reaches a parent table with no explicit SELECT policy"
-    );
-}
-
-#[cfg(feature = "test-utils")]
-#[tokio::test]
-async fn local_insert_with_inherits_policy_allows_missing_parent_policy_in_permissive_local() {
-    let documents_policies = permissions(|p| {
-        p.allow_insert().where_(pe::allowed_to_insert("folder_id"));
-    });
-    let schema = SchemaBuilder::new()
-        .table(TableSchema::builder("folders").column("title", ColumnType::Text))
-        .table(
-            TableSchema::builder("documents")
-                .column("title", ColumnType::Text)
-                .nullable_fk_column("folder_id", "folders")
-                .policies(documents_policies),
-        )
-        .build();
-
-    let client = JazzClient::permissive_test_client(schema).await;
-
-    let folder = client
-        .insert("folders", crate::row_input!("title" => "alice folder"))
-        .expect("seed folder row");
-
-    client
-        .for_session(Session::new("alice"))
-        .insert(
-            "documents",
-            crate::row_input!("title" => "draft doc", "folder_id" => folder.0),
-        )
-    .expect(
-        "permissive local runtimes should treat missing parent INSERT policy as allow for INHERITS",
-    );
-}
-
-#[cfg(feature = "test-utils")]
-#[tokio::test]
-async fn local_update_with_inherits_referencing_allows_missing_source_policy_in_permissive_local() {
-    let files_policies = permissions(|p| {
-        p.allow_update()
-            .where_old(pe::allowed_to_update_referencing("todos", "file_id"))
-            .where_new(pe::always());
-    });
-    let schema = SchemaBuilder::new()
-        .table(
-            TableSchema::builder("files")
-                .column("owner_id", ColumnType::Text)
-                .column("name", ColumnType::Text)
-                .policies(files_policies),
-        )
-        .table(
-            TableSchema::builder("todos")
-                .column("owner_id", ColumnType::Text)
-                .column("title", ColumnType::Text)
-                .nullable_fk_column("file_id", "files"),
-        )
-        .build();
-
-    let client = JazzClient::permissive_test_client(schema).await;
-
-    let file = client
-        .insert(
-            "files",
-            crate::row_input!("owner_id" => "bob", "name" => "shared-file"),
-        )
-        .expect("seed file row");
-    client
-        .insert(
-            "todos",
-            crate::row_input!(
-                "owner_id" => "alice",
-                "title" => "todo referencing file",
-                "file_id" => file.0,
-            ),
-        )
-        .expect("seed referencing todo row");
-
-    client
-        .for_session(Session::new("alice"))
-        .update(
-            file.0,
-            vec![
-                ("owner_id".into(), Value::Text("bob".into())),
-                ("name".into(), Value::Text("updated by alice".into())),
-            ],
-        )
-    .expect(
-        "permissive local runtimes should treat missing source UPDATE policy as allow for INHERITS_REFERENCING",
-    );
-}
-
-#[cfg(feature = "test-utils")]
-#[tokio::test]
-async fn local_update_with_check_inherits_denies_when_parent_is_not_updateable() {
-    let folders_policies = permissions(|p| {
-        p.allow_update()
-            .where_old(pe::eq("owner_id", pe::session("user_id")))
-            .where_new(pe::allowed_to_update_with_depth("parent_id", 10));
-    });
-    let schema = SchemaBuilder::new()
-        .table(
-            TableSchema::builder("folders")
-                .column("owner_id", ColumnType::Text)
-                .column("name", ColumnType::Text)
-                .nullable_fk_column("parent_id", "folders")
-                .policies(folders_policies),
-        )
-        .build();
-    let client = JazzClient::test_client(schema).await;
-
-    let root = client
-        .insert(
-            "folders",
-            crate::row_input!("owner_id" => "alice", "name" => "Root", "parent_id" => Value::Null),
-        )
-        .expect("create root");
-    let child = client
-        .insert(
-            "folders",
-            crate::row_input!("owner_id" => "bob", "name" => "Child", "parent_id" => root.0),
-        )
-        .expect("create child");
-
-    let update_err = client
-        .for_session(Session::new("bob"))
-        .update(
-            child.0,
-            vec![
-                ("owner_id".into(), Value::Text("bob".into())),
-                ("name".into(), Value::Text("Child renamed".into())),
-                ("parent_id".into(), Value::Uuid(root.0)),
-            ],
-        )
-        .expect_err("update should fail inherited WITH CHECK");
-    assert_client_policy_denied(update_err, "folders", Operation::Update);
-}
-
 #[test]
 fn local_update_with_check_inherits_uses_visible_row_region_after_legacy_branch_history_is_removed()
 {
