+++ crates/jazz-tools/tests/policies_integration/declared_fk_inheritance.rs
